ci/requirements.txt
azure-cosmos>=4.9.0
